_Sessions/_2026-01-27_llm-computer-use/poc/requirements.txt
mss>=9.0.1
Pillow>=10.0.0
anthropic>=0.50.0
pyautogui>=0.9.54
